Extract metadata properties from GenericMap into MapMetaMixin
Moves all metadata-derived properties and methods (date handling,
observer coordinate, WCS-adjacent keyword properties, unit parsing,
metadata validation, etc.) out of sunpy/map/mapbase.py into a new
sunpy/map/mixins/mapmeta.py, mixed into GenericMap as MapMetaMixin.

This is the first step of the NDCube-based map refactor, staged as a
pure, behavior-preserving extraction so the eventual base-class swap
is a smaller, more isolated change. PixelPair, SpatialPair, and
MapMetaValidationError move with the mixin but are re-imported into
mapbase.py so existing direct imports from sunpy.map.mapbase (used
throughout sunpy.map.sources and elsewhere) keep working unchanged.

This mixin was originally written by Stuart Mumford, Nabil Freij, and
hayesla on the map_ndcube_migration branch; this commit rebuilds it
against current main.
