+++ sunpy/map/mixins/__init__.py
@@ -1,0 +1,3 @@
+"""
+Mixin classes used to compose `sunpy.map.GenericMap`.
+"""
